+++ cmd/seid/cmd/init.go
@@ -23,8 +23,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/cosmos/cosmos-sdk/x/genutil"
-	"github.com/cosmos/cosmos-sdk/telemetry"
-	serverconfig "github.com/cosmos/cosmos-sdk/server/config"
 )
 
 const (
@@ -144,16 +142,7 @@ func InitCmd(mbm module.BasicManager, defaultNodeHome string) *cobra.Command {
 			toPrint := newPrintInfo(config.Moniker, chainID, nodeID, "", appState)
 
 			tmcfg.WriteConfigFile(filepath.Join(config.RootDir, "config", "config.toml"), config)
-			
-			serverConfig := serverconfig.GetConfig(serverCtx.Viper)
-			if err != nil {
-				return err
-			}
 
-			_, err = startTelemetry(serverConfig)
-			if err != nil {
-				return err
-			}
 
 			return displayInfo(toPrint)
 		},
@@ -168,9 +157,4 @@ func InitCmd(mbm module.BasicManager, defaultNodeHome string) *cobra.Command {
 }
 
 
-func startTelemetry(cfg serverconfig.Config) (*telemetry.Metrics, error) {
-	if !cfg.Telemetry.Enabled {
-		return nil, nil
-	}
-	return telemetry.New(cfg.Telemetry)
-}
+
